backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
anthropic>=0.41.0
python-dotenv>=1.0.1
pydantic>=2.10.6
pydantic-settings>=2.7.1
python-multipart>=0.0.9
httpx>=0.27.0

# Authentication & Database
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# Web Scraping (PJUD)
playwright==1.57.0

# AI Agents (LangGraph + LangChain)
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Document Generation
jinja2>=3.1.0
reportlab>=4.0.0
python-docx>=1.0.0

# Async utilities
nest-asyncio>=1.6.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
